x86/boot/64: Adapt virtual address logic for PIE support
The __startup_64 function assumes all symbols have relocated addresses
instead of the current boot virtual address. PIE generated code favors
relative addresses making all virtual and physical address math incorrect.
To solve this problem, _text is stored in a global and a function is
used to compute relocated addresses which is a no-op when PIE is
disabled.

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>