x86/percpu: Adapt percpu for PIE support
Perpcu uses a clever design where the .percu ELF section has a virtual
address of zero and the custom linux relocation code avoid relocating
specific symbols. It makes the code simple and easily adaptable with or
without SMP support.

This design is incompatible with PIE. While creating a PIE binary, the
compiler tries to make everything relative. The compiler will attempt to
generate instructions with the distance between zero and any 64-bit
virtual address. It will fail as the relocation range cannot fit within
the possible instructions accessing a segment register.

Solve this problem by removing the zero mapping. The .percpu symbols
are now close to the base of the kernel and the compiler generates
appropriate relocations. To accommodate this change, the GS base
is adapted to be the difference between zero and the .percpu section
address. These changes are done only when PIE is enabled. The original
implementation is kept as-is by default.

The assembly and PER_CPU macros are changed to use relative references
when PIE is enabled.

The KALLSYMS_ABSOLUTE_PERCPU configuration is disabled with PIE given
percpu symbols are not absolute in this case.

The stack cookie option is setup with explicit usage of
%gs:fixed_percpu_data+0x28 instead of %fs:0x28 default with PIE. Ensures
the relative percpu can be used with stackprotector.

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>