compiler: Option to default to hidden symbols
Provide an option to change the default visibility to hidden except for
key symbols (like _text). This option is disabled by default and will
be used by x86_64 PIE support to remove errors between compilation units.

The visibility change is also used for external symbols that are
compared as they maybe equals (start/end of sections). In this case,
older versions of GCC will remove the comparison if the symbols are
hidden. This issue exists at least on gcc 4.9 and before.

This change doesn't remove symbols from kallsyms.

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>