parisc: Handle case where flush_cache_range is called with no context
Just when I had decided that flush_cache_range() was always called with
a valid context, Helge reported two cases where the
"BUG_ON(!vma->vm_mm->context);" was hit on the phantom buildd:

[555698.955264] kernel BUG at /mnt/sdb6/linux/linux-4.15.4/arch/parisc/kernel/cache.c:587!
[555699.051275] CPU: 1 PID: 3254 Comm: kworker/1:2 Tainted: G D 4.15.0-1-parisc64-smp #1 Debian 4.15.4-1+b1
[555699.179274] Workqueue: events free_ioctx

[555700.695237]  IAOQ[0]: flush_cache_range+0x164/0x168
[555700.755237]  IAOQ[1]: flush_cache_page+0x0/0x1c8
[555700.811240]  RP(r2): unmap_page_range+0xae8/0xb88
[555700.867231] Backtrace:
[555700.899249]  [<00000000404a6980>] unmap_page_range+0xae8/0xb88
[555700.971238]  [<00000000404a6ae0>] unmap_single_vma+0xc0/0x188
[555701.039238]  [<00000000404a6cdc>] zap_page_range_single+0x134/0x1f8
[555701.115240]  [<00000000404a702c>] unmap_mapping_range+0x1cc/0x208
[555701.191238]  [<0000000040461518>] truncate_pagecache+0x98/0x108
[555701.263239]  [<0000000040461624>] truncate_setsize+0x9c/0xb8
[555701.331263]  [<00000000405d7f30>] put_aio_ring_file+0x80/0x100
[555701.403238]  [<00000000405d803c>] aio_free_ring+0x8c/0x290
[555701.467236]  [<00000000405d82c0>] free_ioctx+0x80/0x180
[555701.531287]  [<0000000040284e6c>] process_one_work+0x21c/0x668
[555701.603237]  [<00000000402854c4>] worker_thread+0x20c/0x778
[555701.671235]  [<0000000040291d44>] kthread+0x2d4/0x2e0
[555701.735235]  [<0000000040204020>] end_fault_vector+0x20/0xc0

This indicates that we need to handle the no context case in
flush_cache_range() as we do in flush_cache_mm().

In thinking about this, I realized that we don't need to flush the TLB
when there is no context.  So, I added context checks to the large flush
cases in flush_cache_mm() and flush_cache_range().  The large flush case
occurs frequently in flush_cache_mm() and the change should improve fork
performance.

The v2 version of this change removes the BUG_ON from flush_cache_page()
by skipping the TLB flush when there is no context.  I also added code
to flush the TLB in flush_cache_mm() and flush_cache_range() when we
have a context that's not current.  Now all three routines handle TLB
flushes in a similar manner.

Signed-off-by: John David Anglin <dave.anglin@bell.net>
Signed-off-by: Helge Deller <deller@gmx.de>